+++ packages/kysely/src/connections/get.ts
@@ -2,6 +2,7 @@ import { Kysely } from "kysely";
 import { removeNullProperties } from "../helpers/remove-nulls";
 import { Connection } from "@authhero/adapter-interfaces";
 import { Database } from "../db";
+import { unflattenObject } from "../flatten";
 
 export function get(db: Kysely<Database>) {
   return async (
@@ -19,6 +20,6 @@ export function get(db: Kysely<Database>) {
       return null;
     }
 
-    return removeNullProperties(connection);
+    return removeNullProperties(unflattenObject(connection, ["options"]));
   };
 }